uint64_t is not size_t on OSX (OSX)

+++ plugins/sparse/sparse_matrix_dataset.cc
@@ -952,7 +952,7 @@ struct SparseMatrixDataset::Itl
         return MR_NO;
     }
 
-    virtual size_t getRowCount() const override
+    virtual uint64_t getRowCount() const override
     {
         auto trans = getReadTransaction();
         return trans->matrix->rowCount();
